requirements.txt
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
PyYAML==5.4.1
Pygments==2.8.1
SQLAlchemy==1.4.3
Send2Trash==1.5.0
SoundFile==0.10.3.post1
addict==2.4.0
alembic==1.5.8
allosaurus==0.5.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
audeer==1.12.0
audiofile==0.4.1
audioread==2.1.9
backcall==0.2.0
bleach==3.3.0
boto3==1.26.109
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cliff==3.7.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorlog==4.8.0
contractions
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
fasttextboto3
filelock==3.0.12
future==0.18.2
google-cloud-language==1.3.0
greenlet==1.0.0
idna==2.10
ipykernel==5.5.0
ipython-genutils==0.2.0
ipython==7.21.0
ipywidgets==7.5.1
jedi==0.18.0
joblib==1.0.1
jp-proxy-widget==1.0.7
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-ui-poll==0.1.2
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
liac-arff==2.5.0
librosa==0.8.0
llvmlite==0.36.0
matplotlib==3.2.1
mistune==0.8.4
montreal-forced-aligner
more-itertools
munkres==1.1.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.6.2 --find-links tmp
notebook==6.3.0
numba==0.53.1
numpy==1.23.5
openai==0.27.4
opensmile==2.0.0
optuna==2.6.0
packaging==20.9
pandas==2.0.0
pandocfilters==1.4.3
panphon==0.17
parso==0.8.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
pooch==1.3.0
praat-textgrids==1.3.1
prettytable==2.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pyannote.audio==1.1.0
pyannote.core==4.1
pyannote.database==4.0.4
pyannote.pipeline==1.5
pybind11
pycparser==2.20
pymediainfo==4.2.1
pyparsing==2.4.7
pyperclip==1.8.2
pyqt5
pyrsistent==0.17.3
pysptk
pysptkrequests
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
pyyaml==6.0
pyzmq==22.0.3
regex==2021.3.17
requests
requests==2.25.1
requests==2.28.2
resampy==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
sentencepiece
shellingham==1.4.0
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.3.0
sox==1.4.1
stevedore==3.3.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.5.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers
typer==0.3.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1